Correct ID for Win32
On the win32 platform ID was coming as the fully qualified path to the
file which caused issues both adding the cover image and retreiving the
epub file from books.json.

We now split the ID on \ and take the last element which should be the
name of the key we need for the database.